* banco2.json: * UnitTest1.cs: * responseProfitSharing.json: * README.md: * response_1621521678296.json: * Startup.cs: * Employee.cs: * Participacao.cs: * EmployeeContextSeed.cs: * ProfitSharingRequest.cs: * ProfitSharingRules.cs: * ProfitSharingResponse.cs:
* EmployeesController.cs: Criação de teste unitario para o relatorio

+++ Employees.API/Employees.API/Entities/Employee.cs
@@ -28,7 +28,7 @@ namespace Employees.API.Entities
         public string Cargo { get; set; }
 
         [BsonElement("salario_bruto")]
-        public decimal SalarioBruto { get; set; }
+        public double SalarioBruto { get; set; }
 
         [BsonElement("data_de_admissao")]
         public DateTime DataAdmissao { get; set; }
