+++ Employees.API/Employees.API/Entities/Participacao.cs
@@ -8,7 +8,7 @@ namespace Employees.API.Entities
         }
         public string Matricula { get; set; }
         public string Nome { get; set; }
-        public Decimal ValorDaParticipacao { get; set; }
+        public Double ValorDaParticipacao { get; set; }
 
     }
 }
